+++ tests/ui/structs/field-implied-unsizing-wfcheck.rs
@@ -7,4 +7,5 @@ struct Bar<T>(T);
 
 fn main() {
     let x = Foo { nested: &Bar(4) };
+    //~^ ERROR the size for values of type `(dyn Debug + 'static)` cannot be known at compilation time
 }
